+++ PCA9635.cpp
@@ -51,13 +51,19 @@ bool PCA9635::begin(uint8_t mode1_mask, uint8_t mode2_mask)
 }
 
 
-void PCA9635::configure(uint8_t mode1_mask, uint8_t mode2_mask)
+uint8_t PCA9635::configure(uint8_t mode1_mask, uint8_t mode2_mask)
 {
   _data = 0;
-  _error = 0;
+  _error = PCA9635_OK;
 
-  setMode1(mode1_mask);
-  setMode2(mode2_mask);
+  uint8_t r1 = setMode1(mode1_mask);
+  uint8_t r2 = setMode2(mode2_mask);
+
+  if ((r1 != PCA9635_OK) || (r2 != PCA9635_OK))
+  {
+    return PCA9635_ERROR;
+  }
+  return _error;
 }
 
 
@@ -97,7 +103,7 @@ uint8_t PCA9635::writeN(uint8_t channel, uint8_t* arr, uint8_t count)
 {
   if (channel + count > _channelCount)
   {
-    _error = PCA9635_ERR_WRITE;
+    _error = PCA9635_ERR_CHAN;
     return PCA9635_ERROR;
   }
   uint8_t base = PCA9635_PWM(channel);
@@ -113,7 +119,8 @@ uint8_t PCA9635::writeN(uint8_t channel, uint8_t* arr, uint8_t count)
     _error = PCA9635_ERR_I2C;
     return PCA9635_ERROR;
   }
-  return PCA9635_OK;
+  _error = PCA9635_OK;
+  return _error;
 }
 
 
@@ -125,7 +132,7 @@ uint8_t PCA9635::writeMode(uint8_t reg, uint8_t value)
     return PCA9635_OK;
   }
   _error = PCA9635_ERR_REG;
-  return PCA9635_ERROR;
+  return _error;
 }
 
 
@@ -139,7 +146,7 @@ uint8_t PCA9635::readMode(uint8_t reg)
     return value;
   }
   _error = PCA9635_ERR_REG;
-  return PCA9635_ERROR;
+  return _error;
 }
 
 
@@ -163,7 +170,8 @@ uint8_t PCA9635::setLedDriverMode(uint8_t channel, uint8_t mode)
   uint8_t clrmask = ~(0x03 << shift);
   uint8_t value = (readReg(reg) & clrmask) | setmask;
   writeReg(reg, value);
-  return PCA9635_OK;
+  _error = PCA9635_OK;
+  return _error;
 }
 
 
@@ -177,7 +185,7 @@ uint8_t PCA9635::getLedDriverMode(uint8_t channel)
   }
 
   uint8_t reg = PCA9635_LEDOUT_BASE + (channel >> 2);
-  uint8_t shift = (channel & 0x03) * 2;  // 0,2,4,6 places
+  uint8_t shift = (channel & 0x03) * 2;  //  0, 2, 4, 6 places
   uint8_t value = (readReg(reg) >> shift ) & 0x03;
   return value;
 }
@@ -203,12 +211,16 @@ bool PCA9635::enableSubCall(uint8_t nr)
 {
   if ((nr == 0) || (nr > 3)) return false;
   uint8_t prev = getMode1();
-  uint8_t reg = prev;
-  if (nr == 1)      reg |= PCA9635_MODE1_SUB1;
-  else if (nr == 2) reg |= PCA9635_MODE1_SUB2;
-  else              reg |= PCA9635_MODE1_SUB3;
+  uint8_t mask = prev;
+  if (nr == 1)      mask |= PCA9635_MODE1_SUB1;
+  else if (nr == 2) mask |= PCA9635_MODE1_SUB2;
+  else              mask |= PCA9635_MODE1_SUB3;
   //  only update if changed.
-  if (reg != prev) setMode1(reg);
+  if (mask != prev)
+  {
+    setMode1(mask);
+    //  TODO error handling ...
+  }
   return true;
 }
 
@@ -217,12 +229,16 @@ bool PCA9635::disableSubCall(uint8_t nr)
 {
   if ((nr == 0) || (nr > 3)) return false;
   uint8_t prev = getMode1();
-  uint8_t reg = prev;
-  if (nr == 1)      reg &= ~PCA9635_MODE1_SUB1;
-  else if (nr == 2) reg &= ~PCA9635_MODE1_SUB2;
-  else              reg &= ~PCA9635_MODE1_SUB3;
+  uint8_t mask = prev;
+  if (nr == 1)      mask &= ~PCA9635_MODE1_SUB1;
+  else if (nr == 2) mask &= ~PCA9635_MODE1_SUB2;
+  else              mask &= ~PCA9635_MODE1_SUB3;
   //  only update if changed.
-  if (reg != prev) setMode1(reg);
+  if (mask != prev)
+  {
+    setMode1(mask);
+    //  TODO error handling ...
+  }
   return true;
 }
 
@@ -230,16 +246,20 @@ bool PCA9635::disableSubCall(uint8_t nr)
 bool PCA9635::isEnabledSubCall(uint8_t nr)
 {
   if ((nr == 0) || (nr > 3)) return false;
-  uint8_t reg = getMode1();
-  if (nr == 1) return (reg & PCA9635_MODE1_SUB1) > 0;
-  if (nr == 2) return (reg & PCA9635_MODE1_SUB2) > 0;
-  return (reg & PCA9635_MODE1_SUB3) > 0;
+  uint8_t mask = getMode1();
+  if (nr == 1) return (mask & PCA9635_MODE1_SUB1) > 0;
+  if (nr == 2) return (mask & PCA9635_MODE1_SUB2) > 0;
+  return (mask & PCA9635_MODE1_SUB3) > 0;
 }
 
 
 bool PCA9635::setSubCallAddress(uint8_t nr, uint8_t address)
 {
-  if ((nr == 0) || (nr > 3)) return false;
+  if ((nr == 0) || (nr > 3))
+  {
+    //  _error = ??  TODO
+    return false;
+  }
   writeReg(PCA9635_SUBADR(nr), address);
   return true;
 }
@@ -247,7 +267,11 @@ bool PCA9635::setSubCallAddress(uint8_t nr, uint8_t address)
 
 uint8_t PCA9635::getSubCallAddress(uint8_t nr)
 {
-  if ((nr == 0) || (nr > 3)) return 0;
+  if ((nr == 0) || (nr > 3))
+  {
+    //  _error = ??  TODO
+    return 0;
+  }
   uint8_t address = readReg(PCA9635_SUBADR(nr));
   return address;
 }
@@ -256,9 +280,13 @@ uint8_t PCA9635::getSubCallAddress(uint8_t nr)
 bool PCA9635::enableAllCall()
 {
   uint8_t prev = getMode1();
-  uint8_t reg = prev | PCA9635_MODE1_ALLCALL;
+  uint8_t mask = prev | PCA9635_MODE1_ALLCALL;
   //  only update if changed.
-  if (reg != prev) setMode1(reg);
+  if (mask != prev)
+  {
+    setMode1(mask);
+    //  error handling TODO
+  }
   return true;
 }
 
@@ -266,17 +294,21 @@ bool PCA9635::enableAllCall()
 bool PCA9635::disableAllCall()
 {
   uint8_t prev = getMode1();
-  uint8_t reg = prev & ~PCA9635_MODE1_ALLCALL;
+  uint8_t mask = prev & ~PCA9635_MODE1_ALLCALL;
   //  only update if changed.
-  if (reg != prev) setMode1(reg);
+  if (mask != prev)
+  {
+    setMode1(mask);
+    //  error handling TODO
+  }
   return true;
 }
 
 
 bool PCA9635::isEnabledAllCall()
 {
-  uint8_t reg = getMode1();
-  return reg & PCA9635_MODE1_ALLCALL;
+  uint8_t mask = getMode1();
+  return mask & PCA9635_MODE1_ALLCALL;
 }
 
 
@@ -303,14 +335,15 @@ bool PCA9635::setOutputEnablePin(uint8_t pin)
   if (_OutputEnablePin != 255)
   {
     _OutputEnablePin = pin;
-    pinMode(pin, OUTPUT);
+    pinMode(_OutputEnablePin, OUTPUT);
+    digitalWrite(_OutputEnablePin, LOW);
     return true;
   }
   return false;
 }
 
 
-bool PCA9635::enableOutput(bool on)
+bool PCA9635::setOutputEnable(bool on)
 {
   if (_OutputEnablePin != 255)
   {
@@ -330,6 +363,7 @@ uint8_t PCA9635::getOutputEnable()
   return LOW;
 }
 
+
 /////////////////////////////////////////////////////
 //
 //  PRIVATE
@@ -340,6 +374,7 @@ uint8_t PCA9635::writeReg(uint8_t reg, uint8_t value)
   _wire->write(reg);
   _wire->write(value);
   _error = _wire->endTransmission();
+
   if (_error == 0) _error = PCA9635_OK;
   else _error = PCA9635_ERR_I2C;
   return _error;
@@ -351,6 +386,7 @@ uint8_t PCA9635::readReg(uint8_t reg)
   _wire->beginTransmission(_address);
   _wire->write(reg);
   _error = _wire->endTransmission();
+
   if (_wire->requestFrom(_address, (uint8_t)1) != 1)
   {
     _error = PCA9635_ERROR;
@@ -362,5 +398,5 @@ uint8_t PCA9635::readReg(uint8_t reg)
 }
 
 
-// -- END OF FILE --
+//  -- END OF FILE --